# docs/public/openapi.yaml
openapi: 3.0.3
info:
  title: pici API
  version: 0.1.0
  description: |
    The pici HTTP API. It drives projects, variables/secrets, and workflow
    executions. Day-to-day operations are usually done through the
    [CLI](/guide/cli) (`pici-cli`), but every operation is available over HTTP.

    ## Authentication

    The API token is required (set via `-api-token` or `PICI_API_TOKEN`).
    Send it either as `Authorization: Bearer <token>` or `X-API-Token: <token>`.

    `/health`, `/api/webhooks/*` and `/badge/*` are exempt (webhooks use their
    own signature/token, badges are meant to be embedded publicly).

    ## Path parameters

    `{id}` on project routes accepts either the project **UUID** or its **name**.
    Execution routes are nested under a project and use `{executionID}`, a
    per-project auto-incrementing number.
servers:
  - url: "{scheme}://{host}:{port}"
    description: pici server (editable from Swagger UI)
    variables:
      scheme:
        default: http
        enum: [http, https]
        description: Protocol
      host:
        default: localhost
        description: Host where pici is running
      port:
        default: "8080"
        description: HTTP listen port (`-http-addr`)
security:
  - BearerAuth: []
  - ApiToken: []
...
paths:
...
  /api/projects:
    post:
      tags: [Projects]
      summary: Register a project
      operationId: createProject
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProjectRequest"
      responses:
        "201":
          description: Project created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalError"
    get:
      tags: [Projects]
      summary: List projects
      operationId: listProjects
      responses:
        "200":
          description: List of projects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Project"
        "500":
          $ref: "#/components/responses/InternalError"
...
  /api/projects/{id}/executions/{executionID}:
    parameters:
      - $ref: "#/components/parameters/ProjectId"
      - $ref: "#/components/parameters/ExecutionId"
    get:
      tags: [Executions]
      summary: Get an execution
      operationId: getExecution
      responses:
        "200":
          description: Execution found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Execution"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
...
  /api/projects/{id}/executions/{executionID}/artifacts:
    parameters:
      - $ref: "#/components/parameters/ProjectId"
      - $ref: "#/components/parameters/ExecutionId"
    get:
      tags: [Artifacts]
      summary: List collected artifacts
      operationId: listArtifacts
      responses:
        "200":
          description: List of artifacts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Artifact"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      description: "Send the API token as `Authorization: Bearer <token>`."
    ApiToken:
      type: apiKey
      in: header
      name: X-API-Token
      description: Send the API token in the `X-API-Token` header.
  parameters:
    ProjectId:
      name: id
      in: path
      required: true
      description: Project ID (UUID) or name
      schema:
        type: string
    ExecutionId:
      name: executionID
      in: path
      required: true
      description: Per-project execution number
      schema:
        type: integer
        format: int64
...
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
  schemas:
    Provider:
      type: string
      enum: [github, gitlab, generic]
    AuthType:
      type: string
      enum: [none, token, ssh]
    Status:
      type: string
      enum: [pending, running, success, failed, canceled]
    Trigger:
      type: string
      enum: [manual, webhook, cron, rebuild]
    Source:
      type: string
      enum: [git, snapshot]
      description: Whether the code was cloned (`git`) or uploaded (`snapshot`).
    StepStatus:
      type: string
      enum: [pending, running, success, failed, skipped, canceled]
...
    Error:
      type: object
      properties:
        error:
          type: string
    Project:
      type: object
      required:
        [id, name, repo_url, provider, auth_type, created_at, updated_at]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        repo_url:
          type: string
        provider:
          $ref: "#/components/schemas/Provider"
        auth_type:
          $ref: "#/components/schemas/AuthType"
        auth_user:
          type: string
          nullable: true
        default_branch:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    CreateProjectRequest:
      type: object
      required: [name, repo_url]
      properties:
        name:
          type: string
        repo_url:
          type: string
        provider:
          $ref: "#/components/schemas/Provider"
          description: Auto-detected from `repo_url` when omitted.
        auth_type:
          $ref: "#/components/schemas/AuthType"
          description: Defaults to `token` when `auth_secret` is set, else `none`.
        auth_user:
          type: string
          description: Defaults to `oauth2` (token) / `git` (ssh).
        auth_secret:
          type: string
          description: Token or SSH private key.
        webhook_secret:
          type: string
        default_branch:
          type: string
...
    StepResult:
      type: object
      required: [name, status, exit_code]
      properties:
        name:
          type: string
        status:
          $ref: "#/components/schemas/StepStatus"
        started_at:
          type: string
          format: date-time
          nullable: true
        finished_at:
          type: string
          format: date-time
          nullable: true
        exit_code:
          type: integer
        error:
          type: string
        env:
          type: array
          items:
            type: string
          description: Environment variables the step ran with (secrets masked).
    Execution:
      type: object
      required:
        [
          id,
          project_id,
          workflow,
          ref,
          commit_sha,
          status,
          trigger,
          source,
          created_at,
        ]
      properties:
        id:
          type: integer
          format: int64
        project_id:
          type: string
          format: uuid
        project:
          type: string
        workflow:
          type: string
        ref:
          type: string
        commit_sha:
          type: string
        status:
          $ref: "#/components/schemas/Status"
        trigger:
          $ref: "#/components/schemas/Trigger"
        source:
          $ref: "#/components/schemas/Source"
        steps:
          type: array
          items:
            $ref: "#/components/schemas/StepResult"
        error:
          type: string
        started_at:
          type: string
          format: date-time
          nullable: true
        setup_finished_at:
          type: string
          format: date-time
          nullable: true
        finished_at:
          type: string
          format: date-time
          nullable: true
        created_at:
          type: string
          format: date-time
...
    Artifact:
      type: object
      required: [step, path, size]
      properties:
        step:
          type: string
        path:
          type: string
        size:
          type: integer
          format: int64
...